Added validation and refactoring

+++ Assets/Source/PostgreSQL/Components/SQLDataDeleter.cs
@@ -16,12 +16,18 @@ namespace LearningStuff.PostgreSQL.Components
             _sqlCommandsExecutor = sqlCommandsExecutor ?? throw new ArgumentNullException(nameof(sqlCommandsExecutor));
         }
 
-        public void DeleteData(string databaseName, SQLArgument[] sqlData)
+        public void DeleteData(string databaseName, SQLArgument[] sqlArguments)
         {
+            if (databaseName == null)
+                throw new ArgumentNullException(nameof(databaseName));
+
+            if (sqlArguments == null || sqlArguments.Length == 0)
+                throw new ArgumentNullException(nameof(sqlArguments));
+
             var finalCommandStringBuilder = new StringBuilder();
             finalCommandStringBuilder.Append($"DELETE FROM {databaseName} WHERE ");
 
-            finalCommandStringBuilder.Append(_sqlParametersStringBuilder.BuildParameters(sqlData.Select(data => $"{data.Name} = {data.Value}").ToArray(), " AND "));
+            finalCommandStringBuilder.Append(_sqlParametersStringBuilder.BuildParameters(sqlArguments.Select(data => $"{data.Name} = {data.Value}").ToArray(), " AND "));
             _sqlCommandsExecutor.ExecuteNonQuery(finalCommandStringBuilder.ToString());
         }
     }
